+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 1. Occupation model & migration
 2. Position model & migration
 3. Post model & migration
+
+## v0.2
+1. Post Reporting Structure model & migration
+2. Improved the Post model to add the #inferior_post_reporting_structures, #superior_post_reporting_structures, #inferior_posts, & #superior_posts associations
+3. Improved the Post model to add the ::report! method